refactor(sdk): extract CIUS-RO party mapping from CLI into SDK
Move countyFromAddress, bucharestSectorFromAddress, companyToParty and
mergePartyOverride into sdk/src/ubl/roPartyMapping.ts so non-CLI consumers
(the fastbill SaaS) can reuse them. CLI re-exports keep its public surface
and tests unchanged.
